Deleted unnecessary files and moved contents to more appropriate space.
